Implement requireMinNonce
